Modify: Seperate MyOrders and OrderBook
